ConnectVmEvent works even when we are on disconnected home screen
